requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pandas
pillow
celery
redis
requests
